archive/20250916-122437/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.21.0
transformers>=4.30.0
accelerate>=0.20.0
optimum[openvino]>=1.13.0
openvino>=2023.1.0
opencv-python>=4.8.0
pillow>=9.5.0
numpy>=1.24.0
imageio>=2.31.0
imageio-ffmpeg>=0.4.8
tqdm>=4.65.0
scipy>=1.10.0
matplotlib>=3.7.0
